requirements.txt
streamlit==1.53.0
python-dotenv==1.2.1
langchain==0.3.0
langchain-community==0.3.0
langchain-core==0.3.0
langchain-openai==0.2.0
langchain-text-splitters==0.3.0
faiss-cpu==1.13.2
openai==1.109.1
beautifulsoup4==4.14.3
bs4==0.0.2
tiktoken==0.12.0